code review: default value of concurrency config

+++ exanho/core/actors/configs/concurrency.py
@@ -2,5 +2,5 @@ from ...common.descriptors import Boolean, Integer, String
 from . import ConfigBase
 
 class Concurrency(ConfigBase):
-    degree = Integer(1)
+    degree = Integer(0)
     kind = String('thread')
